requirements.txt
asgiref==3.2.10
astroid==2.4.2
click==7.1.2
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.0.9
django-appconf==1.0.4
django-ckeditor==6.0.0
django-import-export==2.3.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-sitetree==1.16.0
django-view-breadcrumbs==1.0.0
et-xmlfile==1.0.1
Flask==1.1.2
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupPy==1.14
MarkupSafe==1.1.1
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.5
Pillow==7.2.0
pylint==2.5.3
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
toml==0.10.1
twentytab-cked==0.5
Werkzeug==1.0.1
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0

numpy~=1.20.2
pip~=21.1.1
lxml~=4.6.2
mysqlclient~=2.0.2
setuptools~=51.0.0
matplotlib~=3.4.2